Basic physics cleanup to put the scaling back into the named constants
The ship.time_factor value is now available through status, it is the
real honest time factor; it is set to 10, because this seems to be the
closest to the previous behaviour

Arbitrary constants in time step length, velocity application etc. got
removed. Trajectory autoscaling now hopefully works. The engine upgrade
now should give an actual engine boost…

Updated timings for portal and timed messages.

Acceleration cap has not been touched. It is easy to tune the basic game
physics to work fine without it; trajectory is a problem, though.