Don't copy env table around so much.
This allows our completion logic (which lives at the top-level of
session.lua) to see the changes to env.CWD when you change
directories. However, it also means we can't trust env.USER to be
accurate when doing permissions checks, so we should probably fix
that! At some point, I mean.